+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Ability to use timeOffset only when local system is using non-UTC timezone
 
 ## [1.0.0-beta.2]
 ### Added